Fixed wiki anchor bug
Opening the wiki with an anchor (i.e., following a link pointing to an
anchor) did not result in moving to that pointer, as the content was not
shown (tabpane was not changed). So the content to scroll to had an
offsetTop of 0 and the scroll request could not be serverd.

Solution: First show content and afterwards scroll to it (move
tabpane.change up)